# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Api
 * Jira Ticket Briefing API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UploadedFile {
  name: string;
  content: string;
}

export interface BriefInput {
  /** Jira ticket key, Jira URL, or GitHub PR URL (fallback/legacy) */
  input?: string;
  prUrl?: string;
  jiraUrl?: string;
  jiraContext?: string;
  files?: UploadedFile[];
}

export interface BriefSection {
  title: string;
  content: string;
}

export type BriefMode = typeof BriefMode[keyof typeof BriefMode];


export const BriefMode = {
  'ticket-only': 'ticket-only',
  'ticket-and-pr': 'ticket-and-pr',
  'pr-only': 'pr-only',
} as const;

export interface Brief {
  id: number;
  input: string;
  mode: BriefMode;
  title: string;
  sections: BriefSection[];
  /** @nullable */
  jiraKey?: string | null;
  /** @nullable */
  prUrl?: string | null;
  /** @nullable */
  rawJira?: string | null;
  /** @nullable */
  rawPr?: string | null;
  createdAt: string;
}

export type BriefSummaryMode = typeof BriefSummaryMode[keyof typeof BriefSummaryMode];


export const BriefSummaryMode = {
  'ticket-only': 'ticket-only',
  'ticket-and-pr': 'ticket-and-pr',
  'pr-only': 'pr-only',
} as const;

export interface BriefSummary {
  id: number;
  input: string;
  mode: BriefSummaryMode;
  title: string;
  /** @nullable */
  jiraKey?: string | null;
  /** @nullable */
  prUrl?: string | null;
  createdAt: string;
}

export interface ConfigStatus {
  jiraConfigured: boolean;
  githubConfigured: boolean;
  allConfigured: boolean;
  missingVars?: string[];
}

export interface ErrorResponse {
  error: string;
}