# lib/api-zod/src/generated/types/brief.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Api
 * Jira Ticket Briefing API
 * OpenAPI spec version: 0.1.0
 */
import type { BriefMode } from './briefMode';
import type { BriefSection } from './briefSection';

export interface Brief {
  id: number;
  input: string;
  mode: BriefMode;
  title: string;
  sections: BriefSection[];
  /** @nullable */
  jiraKey?: string | null;
  /** @nullable */
  prUrl?: string | null;
  /** @nullable */
  rawJira?: string | null;
  /** @nullable */
  rawPr?: string | null;
  createdAt: Date;
}