# lib/api-zod/src/generated/types/briefInput.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Api
 * Jira Ticket Briefing API
 * OpenAPI spec version: 0.1.0
 */
import type { UploadedFile } from './uploadedFile';

export interface BriefInput {
  /** Jira ticket key, Jira URL, or GitHub PR URL (fallback/legacy) */
  input?: string;
  prUrl?: string;
  jiraUrl?: string;
  jiraContext?: string;
  files?: UploadedFile[];
}