# lib/api-zod/src/generated/types/briefSummary.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Api
 * Jira Ticket Briefing API
 * OpenAPI spec version: 0.1.0
 */
import type { BriefSummaryMode } from './briefSummaryMode';

export interface BriefSummary {
  id: number;
  input: string;
  mode: BriefSummaryMode;
  title: string;
  /** @nullable */
  jiraKey?: string | null;
  /** @nullable */
  prUrl?: string | null;
  createdAt: Date;
}