# lib/api-zod/src/generated/types/uploadedFile.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Api
 * Jira Ticket Briefing API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadedFile {
  name: string;
  content: string;
}